dt: pass dt_driver as parameter to probe function
In order to allow specific subsystems to redefine the probe function and
thus pass additional information (i2c devices for instance) add the
dt_driver which contains the generic driver information such as
".driver". This is going to be used by I2C framework to use a custom i2c
probe function which takes the i2c device as parameter that can then be
used to communicate using the parent controller.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>