core: register spmc_init() with boot_final()
In case of virtualization registers spmc_init() with boot_final()
instead of service_init() to have my_endpoint_id initialized as part of
boot initialization instead of delayed initialization when the OP-TEE
partition is created.

This guarantees that my_endpoint_id holds the correct value when the
first FF-A request is received.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>