drivers: implement qm driver
The Hisilicon QM is a Queue Management module.
In order to unify the interface between accelerator and software,
a unified queue management module QM is used to interact with software.
Each accelerator module integrates a QM. Software issues tasks to the SQ,
and the QM obtains the address of the SQE. The BD information is sent to
the accelerator. After the task processing is complete, the accelerator
applies for a write-back address from the QM to write back the SQ

Signed-off-by: Xiaoxu Zeng <zengxiaoxu@huawei.com>