core: mm: Fix idx truncation bug
When the idx of TA memory mapping is assigned, it is saved in the
l1_idx_t, which is the uint8_t or uint16_t type. But when it is
parsed, it is save in the only uint8_t that will lead to error.
To solve this problem, the idx shouldbe saved in the uint16_t
type when parsing the idx.

Signed-off-by: leisen <leisen1@huawei.com>