crypto: ecc: remove the SM2 encrypt/decrypt formatting from the crypto driver
Remove the SM2 formatting for encrypt and decrypt operations as it is
incompatible with the software crypto fallback mechanism.

When using the software fallback mechanism, the hardware crypto driver
behind the crypto driver API would eventually call the software crypto
implementation if the given algorithm is not supported by the hardware.

In that case, for SM2 encryption, the data would be formatted by the
crypto driver API before calling the software crypto libraries. The
crypto software libraries expect raw data as input and not pre-formatted.

The data formatting should be handled by the hardware crypto driver.

Signed-off-by: Clement Faure <clement.faure@nxp.com>