core: console: allow use of multiple consoles
This allows registering multiple serial consoles The main use for this
would be with the new cbmem console driver along with a serial console.

If a driver is re-registered, then it is a no-op. But if more than one
driver is registered, then a linked list of drivers is used to perform
logging across all of them.

Signed-off-by: Jeffrey Kardatzke <jkardatzke@google.com>