core: add external device tree support
Add support for an external repository for device tree source files.

To manage device-tree compilation, the dts folder is given as an include
argument of the dts pre-processing command.

Implement support of dts-source-path makefile variable that should
be overridden at platform level with the external device tree source
file path.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>