drivers: stm32_gpio: helper function to register GPIO banks
Adds an helper function (dt_stm32_gpio_bank()) to register GPIO banks
in stm32_gpio driver based on DT pinctrl nodes. GPIO banks are
registered in a local list from which will be later used to find a
bank based on its ID number. The function is expected to be called
from the driver probe sequence. The function ensures a bank is not
registered twice.

Acked-by: Lionel Debieve <lionel.debieve@foss.st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>