ffa: remove ffa.h and smccc.h from host export list
commit 5489e94f0566 ("core: ffa: add boot info structs and defines")
causes the compiler error below in host applications.

ffa.h:196:41: error: implicit declaration of function ‘U’ [-Werror=implicit-function-declaration]
  196 | #define FFA_BOOT_INFO_NAME_LEN          U(16)
      |                                         ^
ffa.h:319:19: note: in expansion of macro ‘FFA_BOOT_INFO_NAME_LEN’
  319 |         char name[FFA_BOOT_INFO_NAME_LEN];
      |                   ^~~~~~~~~~~~~~~~~~~~~~
ffa.h:319:14: error: variably modified ‘name’ at file scope
  319 |         char name[FFA_BOOT_INFO_NAME_LEN];
      |              ^~~~

ffa.h uses the U macro which is only part of OP-TEE's libc and not a
standard macro. This causes a compiler error when compiling a host
application and the libc is provided by the compiler. This makes ffa.h
not suitable for host export.

This patch resolves the issue by removing ffa.h from the host export
list to prevent external components to use it. TAs and client
applications have no use of this file.

Also remove smccc.h as it was only added to the list because it is a
dependency of ffa.h

Signed-off-by: Imre Kis <imre.kis@arm.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>