ci.yml: Temporarily don't build libscmi-server
PR812 in SCP-firmware [1] has added some source files into
mod_scmi_perf directory, but it currently fails when building for the
optee-os case.
This is due to the fact that the current build system in optee, for
building the scmi-server fetched from SCP-firmware codebase, works well
under the assumption that only one source file is required.

To unlock the circular dependency between build systems and source
codes, temporarily remove the build for the scmi-server.

This will allow the following two patches to be safely merged into the
optee_os repo:
core: scmi-server: Add support for conditional options for SCMI-Perf
optee-fvp: Enable CFG_SCPFW_SCMI_PERF_PROTOCOL_OPS

Subsequently, PR812 in SCP-firmware can be merged, and immediately
a revert patch for this present one will follow to reintroduce the
libscmi-server build.

Link: https://github.com/ARM-software/SCP-firmware/pull/812 [1]
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>